refactor(skins): separate source and build transforms
