+++ packages/skins/src/components/sliders/slider.tsx
@@ -1,5 +1,5 @@
 import type { SkinComponentMeta } from '../../meta';
-import { defineRenderTarget } from '../../render-target';
+import { defineRenderTarget } from '../../render';
 import styles from '../../styles/sliders/slider.styles';
 
 /** Shared slider track. */
